move all logics to controllers
